requirements.txt
rdkit
pandas
numpy
torch
scipy
scikit-learn
jax
jaxlib
jraph
dm-haiku
optax